The sparc64 compiler is a casualty of this merge, mostly because
of 8e7a3ea (cmd/compile/internal: named types for Etype and Op in
struct Node).  It will be added in a subsequent commit.

 Conflicts:
	src/cmd/asm/internal/arch/arch.go
	src/cmd/asm/internal/asm/asm.go
	src/cmd/compile/internal/gc/cgen.go
	src/cmd/compile/internal/gc/gsubr.go
	src/cmd/compile/internal/gc/pgen.go
	src/cmd/compile/internal/gc/reg.go
	src/cmd/compile/internal/gc/walk.go
	src/cmd/dist/buildtool.go
	src/cmd/internal/obj/link.go
	src/cmd/internal/obj/util.go
	src/cmd/link/internal/ld/arch.go
	src/cmd/link/internal/ld/dwarf.go
	src/cmd/link/internal/ld/elf.go
	src/cmd/link/internal/ld/ldelf.go
	src/cmd/link/internal/ld/symtab.go

+++ src/cmd/compile/internal/arm64/cgen.go
@@ -48,7 +48,7 @@ func blockcopy(n, res *gc.Node, osrc, odst, w int64) {
 	// the src and dst overlap, then reverse direction
 	dir := align
 
-	if osrc < odst && int64(odst) < int64(osrc)+w {
+	if osrc < odst && odst < osrc+w {
 		dir = -dir
 	}
 
@@ -129,7 +129,7 @@ func blockcopy(n, res *gc.Node, osrc, odst, w int64) {
 		// TODO(austin): Instead of generating ADD $-8,R8; ADD
 		// $-8,R7; n*(MOVDU 8(R8),R9; MOVDU R9,8(R7);) just
 		// generate the offsets directly and eliminate the
-		// ADDs.  That will produce shorter, more
+		// ADDs. That will produce shorter, more
 		// pipeline-able code.
 		var p *obj.Prog
 		for ; c > 0; c-- {
